Refactor classes in packages
