+++ build.gradle
@@ -40,6 +40,8 @@ shadowJar {
 checkstyle {
     toolVersion = '8.23'
 }
+
 javafx {
     version = "11.0.2"
-    modules = [ 'javafx.controls', 'javafx.fxml' ]
+    modules = ['javafx.controls', 'javafx.fxml']
+}
