+++ src/main/java/Main.java
@@ -12,7 +12,8 @@ import javafx.stage.Stage;
  */
 public class Main extends Application {
 
-    private Duke duke = new Duke("list.txt");
+//    private Duke duke = new Duke("list.txt");
+    private Duke duke = new Duke();
 
     @Override
     public void start(Stage stage) {
