+++ src/main/java/duke/DukeException.java
@@ -1,5 +1,7 @@
+package duke;
+
 // Custom exception class for Duke
-class DukeException extends Exception {
+public class DukeException extends Exception {
     public DukeException(String errorMessage) {
         super(errorMessage);
     }
