+++ src/main/java/task/ToDo.java
@@ -1,4 +1,6 @@
-class ToDo extends Task {
+package task;
+
+public class ToDo extends Task {
 
     public ToDo(String description) {
         super(description);
@@ -11,6 +13,6 @@ class ToDo extends Task {
 
     @Override
     public String toSave() {
-        return "T | " + (isDone ? "1" : "0") + " | " + this.description;
+        return "T | " + (getStatus() ? "1" : "0") + " | " + getDescription();
     }
 }
