+++ src/test/java/DeadlineTest.java
@@ -1,4 +1,6 @@
 import org.junit.jupiter.api.Test;
+import task.Deadline;
+import task.Task;
 
 import static org.junit.jupiter.api.Assertions.*;
