+++ src/test/java/EventTest.java
@@ -1,4 +1,6 @@
 import org.junit.jupiter.api.Test;
+import task.Event;
+import task.Task;
 
 import static org.junit.jupiter.api.Assertions.*;
