+++ src/test/java/ParserTest.java
@@ -1,3 +1,5 @@
+import duke.DukeException;
+import command.Parser;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
