+++ src/test/java/TaskListTest.java
@@ -1,4 +1,5 @@
 import org.junit.jupiter.api.Test;
+import task.*;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
